+++ CHANGELOG.md
@@ -3,3 +3,8 @@
 - TextField: Updated the order of the attachments
 - trapFocus: Fixed aria-hidden not removing for some of the event handlers combinations
 - Carousel: Fixed scroll snapping scrolling the carousel back
+- TextField: Rounded flag
+
+Documentation
+
+- Included next/font integration for next15